Check the maximum allowed accuracy level even for system applications
Respect the value of the user preference concerning the usage of their
geolocation. This should fix the privacy and GDPR conformity concerns as
the user explicitly needs to enable the option.

Note that there is no distinction between the system applications
anymore, turning on the option is allowing them all to get the location.

Closes: #924516 #958497
Gbp-Dch: Full